requirements.txt
Django==3.1
djangorestframework==3.12.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.8.0
drf-yasg==1.21.5
django-filter==2.4.0
xlwt==1.3.0
openpyxl==3.0.9
rsa==4.9
xlrd==1.2.0
pandas==1.3.5
django_extensions==2.1.9